Fix casts for 64bit osx, update changes to 3.0.2 on osx

+++ AgilityBook/src/Win/DlgConfigEventMethod.cpp
@@ -560,7 +560,7 @@ DEFINE_ON_INIT(CDlgConfigEventMethod)
 
 ARBConfigScoring::ScoringStyle CDlgConfigEventMethod::GetType(int index) const
 {
-	return static_cast<ARBConfigScoring::ScoringStyle>(reinterpret_cast<int>(m_ctrlType->GetClientData(index)));
+	return static_cast<ARBConfigScoring::ScoringStyle>(reinterpret_cast<size_t>(m_ctrlType->GetClientData(index)));
 }
